+++ src/App.css
@@ -1,5 +1,4 @@
 .App {
-  text-align: center;
   display: flex;
   justify-content: space-around;
 }
