+++ src/components/editor/nodes.ts
@@ -1,0 +1,16 @@
+import { HeadingNode, QuoteNode } from "@lexical/rich-text";
+import { ListItemNode, ListNode } from "@lexical/list";
+import { CodeHighlightNode, CodeNode } from "@lexical/code";
+import { AutoLinkNode, LinkNode } from "@lexical/link";
+import { Klass, LexicalNode } from "lexical";
+
+export const nodes: Klass<LexicalNode>[] = [
+  HeadingNode,
+  QuoteNode,
+  ListItemNode,
+  ListNode,
+  CodeNode,
+  CodeHighlightNode,
+  AutoLinkNode,
+  LinkNode,
+];
