+++ src/components/editor/plugins/MarkdownPlugin.tsx
@@ -1,0 +1,6 @@
+import { TRANSFORMERS } from "@lexical/markdown";
+import { MarkdownShortcutPlugin } from "@lexical/react/LexicalMarkdownShortcutPlugin";
+
+export const MarkdownPlugin: React.FC = () => {
+  return <MarkdownShortcutPlugin transformers={TRANSFORMERS} />;
+};
